requirements.txt
openai
anthropic
python-dotenv
pandas
requests
beautifulsoup4
chromadb
streamlit
schedule
matplotlib
